SPX-8140 - Move styles to the top level if after internal direct changes they apeared on the level below
This issue was seen on Osx when autocorrection makes some changes. It adds font tag and if there were some styles then can appear on different levels ( some on font, some on upper span level). It causes issues with line height and highlighting. This fix tries to keep styles on the top level.