SPI-13079 - Improve applying format to collapsed selection
Previously placeholder element with invisible whitespace ("Cursor") was added to caret position and format was applied to it. It caused issues with suggestions and autocapitalization on iOS ( it saw some char and thought that it is the first char of the word). New implementation remembers applied format and applies it when entering the first char.