Change safari runner
There are issues with the cache when running tests in Safari. Sometimes it seems to use cached state and runs tests for some old state of the repo.
The way to clean the caches which I found is to use applescript to select menu item in Safari. As it is already applescript seems like it would be easier to use applescript runner.
I've forked it and used here.