SPX-7801 - Another approach to work with header tags ( H1..H6)
Header tags modify active font size, so when no size is specified it won't be default(14 or 16), but something else. Also as a block level tag it will change behaviour (for example adding newlines).
To simplify things for us, we convert header tags into style ones with corresponding font-size/bold